requirements.txt
scipy==1.6.0
Pillow==8.1.0
numpy==1.19.3
